… bdi_kernel/docs/PHASE6_BUILD_OPTIMIZATION.md
… bdi_kernel/kernel/autoprofiler.h
… bdi_kernel/kernel/optimization.h
… build_config.mk
… linker.ld